requirements.txt
# Django
psycopg2-binary
djangorestframework
markdown
django-filter
django-rest-knox
Pillow
django-cors-headers
pylint-django
# gunicorn==20.0.4
Django==3.2.4
uWSGI
pandas
# >=2.0.18,<2.1